% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree_color.R
\name{get_default_tree_color}
\alias{get_default_tree_color}
\title{Get the default color for a tree}
\usage{
get_default_tree_color(crown_type)
}
\arguments{
\item{crown_type}{Crown type. Full list of options:
\code{"columnar"}
\code{"pyramidal1"}
\code{"pyramidal2"}
\code{"oval"}
\code{"palm"}
\code{"rounded"}
\code{"spreading1"}
\code{"spreading2"}
\code{"vase"}
\code{"weeping"}}
}
\value{
Hex color
}
\description{
Get the default color for a tree
}
\keyword{internal}
