% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_data.R
\name{get_tree_data}
\alias{get_tree_data}
\title{Get Tree Data}
\usage{
get_tree_data(tree_name = NULL)
}
\arguments{
\item{name}{Default \code{NULL}. The tree name, if only a single tree's info is needed. Otherwise,
all trees are returned.}
}
\value{
Returns a data frame.
}
\description{
This function retrieves a data frame of tree default data.
}
\examples{
# Fetch the tree data frame.
get_tree_data()

# Get single tree
get_tree_data("palm")
}
