% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trunk_file.R
\name{get_trunk_file}
\alias{get_trunk_file}
\title{Get Trunk File}
\usage{
get_trunk_file(scaled = TRUE, offset = TRUE)
}
\arguments{
\item{scaled}{Default \code{TRUE}. This parameter determines whether the trunk model should be scaled from a radius of 0.1 to 1. If TRUE, the trunk model will be scaled; otherwise, it will use the original size.}

\item{offset}{Default \code{TRUE}. This parameter determines whether the trunk model should be offset to start at the origin, or centered at the origin (\code{FALSE}).}
}
\value{
Returns the filename of the selected OBJ file (as a .txt file) for the tree trunk. This file can be used to create tree trunk models in a 3D scene.
}
\description{
This function retrieves a trunk OBJ file. The trunk OBJ file can be used to create a tree trunk in a 3D scene. The function has an option to scale the trunk model based on the radius.
}
\examples{
# Fetch an trunk OBJ path
get_trunk_file()
}
