% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_mesh.R
\name{tree_mesh}
\alias{tree_mesh}
\title{Tree Mesh}
\usage{
tree_mesh(
  crown_type = "oval",
  position = c(0, 0, 0),
  angle = 0,
  solid = FALSE,
  resolution = "medium",
  filename = NULL,
  crown_height = 1,
  crown_diameter = 1,
  trunk_height = 1,
  trunk_diameter = 0.1,
  crown_color = "darkgreen",
  trunk_color = "#725c42",
  diffuse_intensity = 1,
  ambient_intensity = 0.2
)
}
\arguments{
\item{crown_type}{Default \code{"oval"}. Crown type. Full list of options
(star indicates both flat and solid variants available, no star indicates only flat version available):\tabular{cc}{
   \strong{3D/Flat} \tab \strong{Name} \cr
   Both \tab \code{"columnar"} \cr
   Both \tab \code{"evergreen1"} \cr
   Both \tab \code{"evergreen2"} \cr
   Both \tab \code{"weeping"} \cr
   Both \tab \code{"spreading1"} \cr
   Both \tab \code{"oval"} \cr
   Flat-only \tab \code{"palm"} \cr
   Flat-only \tab \code{"rounded"} \cr
   Flat-only \tab \code{"spreading2"} \cr
   Flat-only \tab \code{"vase"} \cr
}}

\item{position}{Default \code{c(0,0,0)}. X/Y/Z position of the mesh.}

\item{angle}{Default \code{0}. Amount of rotation around the y-axis for the tree.}

\item{solid}{Default \code{FALSE}. Whether the crown should be a solid mesh, or a collection of flat
2D planes.}

\item{resolution}{Default \code{"medium"}. Level of detail of the tree mesh. All options:
\code{"low"}
\code{"medium"}
\verb{"high}}

\item{filename}{Default \code{NULL}. Filename of the OBJ file, if saving the mesh to a local file.}

\item{crown_height}{Default \code{1}. Crown height.}

\item{crown_diameter}{Default \code{1}. Crown diameter.}

\item{trunk_height}{Default \code{1}. Trunk height.}

\item{trunk_diameter}{Default \code{0.1}. Trunk diameter.}

\item{crown_color}{Default \code{0.1}. Crown color.}

\item{trunk_color}{Default \code{0.1}. Trunk color.}

\item{diffuse_intensity}{Default \code{1.0}. Amount of diffuse (shaded) color included in the model.}

\item{ambient_intensity}{Default \code{0.2}. Amount of ambient (constant) color included in the model.}
}
\value{
\code{ray_mesh} list object
}
\description{
Loads an tree OBJ file as a \code{ray_mesh} object.
}
\examples{
#Load a tree
}
