% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree_data.R
\name{get_tree_data}
\alias{get_tree_data}
\title{Get Tree Data}
\usage{
get_tree_data()
}
\value{
Returns a data frame.
}
\description{
This function retrieves a data frame of tree default data.
}
\examples{
# Fetch the tree data frame.
get_tree_data()
}
