% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trunk_file.R
\name{get_trunk_file}
\alias{get_trunk_file}
\title{Get Trunk File}
\usage{
get_trunk_file()
}
\value{
Filename of OBJ file (as \code{.txt} file)
}
\description{
Description
}
\examples{
#Load an arrow OBJ
get_trunk_file()
}
